feat(dhcp): SOLICIT / ADVERTISE is a thing now...
Yipee. It works...

The design ended up having two actors. One for reading from the stream
and one for writing. The writing actor ends up communicating with the
reading one, so in practice there's a:

user handle -> writing actor -> reading actor

hierarchy to it all, which makes sense since the writing actor wants to
wait for a reply after sending a message first, and bubble that up to
the user. In theory, the DHCP server can also send reconfigure which I
haven't decided how to handle yet though. Probably, the reader will
also hold a handle to the writing actor, but that creates a cycle
between them. Unsure how closing these are gonna work, the cycle might
mean they don't close.

Apologies for any reader. I'm sleep deprived currently and this is just
a brain splurge for a fairly early version of this project